Path sort and code refactor MT-1
