+++ src/App.tsx
@@ -1,18 +1,15 @@
+import { Theme, ThemeProvider } from '@emotion/react';
 import { AppLayout } from '@components/AppLayout';
-import { ThemeProvider } from '@emotion/react';
 import { MainPage } from '@pages/MainPage';
 import { theme } from '@theme/index';
-import { Theme } from '@emotion/react';
 
 const App = () => {
   return (
-    <>
-      <ThemeProvider theme={theme as Theme}>
-        <AppLayout>
-          <MainPage />
-        </AppLayout>
-      </ThemeProvider>
-    </>
+    <ThemeProvider theme={theme as Theme}>
+      <AppLayout>
+        <MainPage />
+      </AppLayout>
+    </ThemeProvider>
   );
 };
